Create cookie clicker cheat page
this isn't really a hack, it's an in-game feature, but here's how to do it. go to your username (e.g. i love cookies's bakery) and paste in this. just delete the brackets and write your own username. 

credits to a kid in my school whos name I can't tell you.